Added file output writting of result

+++ Project/Code/Nearest Neighbor/struct.h
@@ -2,6 +2,6 @@ struct Nodes {
    int x;
    int y;
    double *neighbourDistance;
-   double *neighbourID;
+   int *neighbourID;
  
 };
